CREDIT CARD FRAUD DETECTION PROJECT
Credit card fraud is a major problem that affects financial institutions and cardholders worldwide. According to a report by the Federal Trade Commission, credit card fraud was the most common form of identity theft reported by consumers in the United States. In this project, we aim to build a machine learning model that can detect fraudulent credit card transactions with high accuracy.